Optimize kernels
Use texture border addressing to eliminate some conditional branching
Eliminate size_t from device code to improve register usage

+++ DoseCUDA/dose_kernels/IMRTClasses.cuh
@@ -22,12 +22,12 @@ class IMRTBeam : public CudaBeam{
 
         struct Model {
 
-            size_t n_profile_points;
+            int n_profile_points;
             float * profile_radius;
             float * profile_intensities;
             float * profile_softening;
 
-            size_t n_spectral_energies;
+            int n_spectral_energies;
             float * spectrum_attenuation_coefficients;
             float * spectrum_primary_weights;
             float * spectrum_scatter_weights;
